# docs/public/openapi.yaml
openapi: 3.1.0
info:
  title: BetterStatusPage admin API
  version: 1.0.0
  description: |
    The part of the admin API meant for scripts, CI/CD pipelines and infrastructure tooling. It is the same API the
    admin panel uses, limited to what an API token may call; see https://docs.betterstatuspage.dev/api/.

    **Authentication.** Send an API token (Administration → API tokens) as `Authorization: Bearer bsp_...`. A token
    carries permissions per part of the application: `monitors`, `channels`, `incidents`, `maintenance`,
    `subscribers`, `appearance` (layout and branding) and `reports`, `audit`, `system` (read only), each `read` or
    `write` (write includes read). The separate `vault:use` permission lets a token attach vault secrets to monitors
    and channels; without it a token cannot add, change or move a vault reference. Each operation below names the
    permission it needs in `x-scope`. A token expires after 90 days unless it was created without an expiry.

    **Secrets.** Passwords, client secrets, credential headers, Slack, Discord and Teams webhook URLs and bot tokens in a `config` are never
    returned: they read back as `••••••••`. Send that value back unchanged to keep the stored secret, or send a new
    value (text) to replace it. A masked value that does not match the stored secret, or one that looks like a
    half-edited mask, is refused with `400`. A kept secret must stay where it is sent: if the `url` (or the OAuth2
...

...

...
servers:
...
    variables:
...
security:
  - bearerAuth: []
...
paths:
...
  /maintenance:
    get:
      tags: [Maintenance]
      summary: List maintenance windows
      x-scope: maintenance:read
      responses:
        '200':
          description: Every window.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/MaintenanceWindow' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Maintenance]
      summary: Schedule a maintenance window
      description: Subscribers are told about a window that has not ended yet, unless `notifySubscribers` is false.
      x-scope: maintenance:write
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MaintenanceCreate' }
      responses:
        '200':
          description: The created window.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MaintenanceWindow' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /maintenance/active:
    get:
      tags: [Maintenance]
      summary: List the windows that are in progress now
      x-scope: maintenance:read
      responses:
        '200':
          description: Windows whose start has passed and whose end has not.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/MaintenanceWindow' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /maintenance/{id}:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    get:
      tags: [Maintenance]
      summary: Get a maintenance window
      x-scope: maintenance:read
      responses:
        '200':
          description: The window.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MaintenanceWindow' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Maintenance]
      summary: Change a maintenance window
      description: Only the fields you send change. `monitorIds` replaces the whole list.
      x-scope: maintenance:write
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MaintenancePatch' }
      responses:
        '200':
          description: The updated window.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MaintenanceWindow' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Maintenance]
      summary: Delete a maintenance window
      x-scope: maintenance:write
      responses:
        '204': { description: Deleted. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: An API token created in Administration → API tokens.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer }
  responses:
    BadRequest:
      description: The request is not valid.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing, unknown, expired or revoked token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: The token does not have the permission this operation needs (the message names it), or the user's role may not do it.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: No such object.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    MaintenanceWindow:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        startsAt: { type: integer }
        endsAt: { type: integer }
        description: { type: [string, 'null'] }
        monitorIds:
          type: array
          items: { type: integer }
        createdAt: { type: integer }
        updatedAt: { type: integer }
    MaintenanceCreate:
      type: object
      required: [name, startsAt, endsAt]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        startsAt: { type: integer, description: Milliseconds since the epoch. }
        endsAt: { type: integer, description: Milliseconds since the epoch; must be after `startsAt`. }
        description: { type: [string, 'null'] }
        monitorIds:
          type: array
          items: { type: integer }
        notifySubscribers: { type: boolean, default: true }
    MaintenancePatch:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        startsAt: { type: integer }
        endsAt: { type: integer }
        description: { type: [string, 'null'] }
        monitorIds:
          type: array
          items: { type: integer }
...